pyproject.toml
[tool.poetry]
name = "minecraft-llm-agent-community"
version = "0.1.0"
description = ""
authors = ["naem1023 <relilau00@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.9"
tqdm = "^4.66.4"
langchain = "^0.2.0"
javascript = "^1!1.2.0"
openai = "^1.30.1"
chardet = "^5.2.0"
cchardet = "^2.1.7"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
requests = "^2.32.2"
gymnasium = "^0.29.1"
psutil = "^5.9.8"
minecraft-launcher-lib = "^6.5"
langchain-community = "^0.2.1"
langchain-openai = "^0.1.7"
python-dotenv = "^1.0.1"
langchain-chroma = "^0.1.1"
logzero = "^1.7.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
